requirements.txt
fastapi==0.98.0
uvicorn==0.23.1
sqlalchemy==2.0.19
pymysql==1.0.3
pymongo==4.7.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
python-jose==3.3.0
sentence-transformers==2.2.2
